% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{circular_hough_transform}
\alias{circular_hough_transform}
\alias{compile_circular_hough}
\title{Calculates a circular hough transform}
\usage{
circular_hough_transform(img, width, sobel_width, tolerance = 11)

compile_circular_hough(width, sobel_width, dim_img, tolerance = 11)
}
\arguments{
\item{img}{image to transform}

\item{width}{expected width of the circles}

\item{sobel_width}{width of the sobel filter that will be used to detect the edges}
}
\value{
image
}
\description{
This transform accentuates circular shapes
}
\section{Functions}{
\itemize{
\item \code{compile_circular_hough}: pre-calculates parameters and returns a function
}}

\examples{
x = readImage(system.file('images', 'shapes.png', package='EBImage'))
hough <- circular_hough_transform(x, 51, 3, 5)
# display(normalize(hough))
}
\seealso{
Other filters: 
\code{\link{compile_hough_circle_draw}()},
\code{\link{compile_sobel_filter}()},
\code{\link{compile_variance_filter}()},
\code{\link{dog_filter}()}
}
\concept{filters}
