% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superpixels.R
\name{classify_img_sliced}
\alias{classify_img_sliced}
\title{Classify images as superpixel slices}
\usage{
classify_img_sliced(
  classifier,
  img,
  slices = slic(sobel_filter(img, 5))[[2]],
  features,
  class_highlight,
  ...
)
}
\arguments{
\item{classifier}{a classifier}

\item{img}{image to classify}

\item{features}{a features frame to be used for the classification}

\item{class_highlight}{a classiication to highlight if the prediction is categorical}

\item{...}{additional arguments passed to kmeans}
}
\value{

}
\description{
Classify images as superpixel slices
}
\examples{
# trainset <- build_train_multi(
# params_df,
# train_size_each = 5000,
# filter_widths = c(3,5))
#
# classifier <- ranger::ranger(pixel_class ~ ., trainset)
# features <- calc_features(base_img, c(3,5))
# sliced_img <- slic(sobel_filter(base_img, 5))
# classify_img_sliced(classifier, base_img, sliced_img[[2]],features, "sphere")
}
\seealso{
slic
}
