% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_image_border}
\alias{clean_image_border}
\title{Makes all the values in the edges of a matrix or image a value}
\usage{
clean_image_border(img, clean_size, val = 0)
}
\arguments{
\item{img}{image to use}

\item{clean_size}{number of rows or cols to remove on each edge}

\item{val}{value to be used to replace 0}
}
\value{

}
\description{
Makes all the values in the edges of a matrix or image a value
}
\examples{
x = readImage(system.file('images', 'shapes.png', package='EBImage'))
sobel <- sobel_filter(x, 15)
display(sobel, "raster")
display(clean_image_border(sobel, 50, 0))
}
