% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{display_filters}
\alias{display_filters}
\title{Displays the data from a feature frame}
\usage{
display_filters(feature_df, dims)
}
\arguments{
\item{dims}{}
}
\value{

}
\description{
Displays the data from a feature frame
}
\examples{
myfile <- system.file("extdata", "4T1-shNT-1.png", package = "clasifierrr")
myimg <- readImageBw(myfile)
myfeat <- calc_features(myimg, c(3,5))
display_filters(myfeat, dim(myimg))
}
