% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{dog_filter}
\alias{dog_filter}
\alias{compile_dog_filter}
\title{Calculates the difference of gaussian filter}
\usage{
dog_filter(img, width, ratio = 5, img_fft = fftwtools::fftw2d(img))

compile_dog_filter(width, dim_img, ratio = 5)
}
\arguments{
\item{img}{Dimensions of the image}

\item{width}{width of the filter to be applied}

\item{ratio}{the ratio of the sigma between the blurrier and
the less blurry images}

\item{img_fft}{optional argument to pass the pre-computed fourier tranform of
the image from `fftwtools::fftw2d(img)`.}

\item{dim_img}{the dimensions of the image to be used later during filtering,
only needed if pre-compiling the filter.}
}
\value{

}
\description{
This filter attempts to enhance the features in an image, it is based on
subtracting a blurry image from a less blurry one. The ratio of this
blurrines is measured by the ratio in their `sigma`.
}
\section{Functions}{
\itemize{
\item \code{compile_dog_filter}: pre-calculates parameters and returns a function
}}

\examples{
test_image <- matrix(runif(81), 9)
dog_filter(test_image, 3)

my_dog_filter <- compile_dog_filter(3, dim(test_image))
my_dog_filter(test_image)
}
\seealso{
Other filters: 
\code{\link{circular_hough_transform}()},
\code{\link{compile_hough_circle_draw}()},
\code{\link{compile_sobel_filter}()},
\code{\link{compile_variance_filter}()}
}
\concept{filters}
