% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{filter_masks}
\alias{filter_masks}
\title{Filters small or large objects}
\usage{
filter_masks(mask, min_radius, max_radius)
}
\arguments{
\item{mask}{an image object of type mask}

\item{min_radius}{minimum radius of objects to keep}

\item{max_radius}{maximum radius of objects to keep}
}
\value{

}
\description{
Filters out in a mask objects that are too little or too large
}
\examples{
img <- readImageBw(system.file("extdata", "tiny_4T1-shNT-1.png", package = "clasifierrr"))
mask <- EBImage::thresh(img)

# EBImage::display(mask)

# EBImage::display(filter_masks(mask, min_radius = 2, max_radius = Inf))
# EBImage::display(filter_masks(mask, min_radius = 0, max_radius = 20))

}
