% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_edges}
\alias{generate_edges}
\title{Generate an edge image of a mask, based on an image}
\usage{
generate_edges(img, mask, opac = c(1, 0.8), cols = c("#FDE725FF", "#440154FF"))
}
\arguments{
\item{img}{image}

\item{mask}{mask}

\item{opac}{}
}
\value{

}
\description{
Generate an edge image of a mask, based on an image
}
