% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_overlay}
\alias{generate_overlay}
\title{Generate an overlaid image of a mask and a base image}
\usage{
generate_overlay(
  img,
  mask,
  opac = c(1, 0.8),
  cols = c("#FDE725FF", "#440154FF")
)
}
\arguments{
\item{img}{image}

\item{mask}{mask}

\item{opac}{}
}
\value{

}
\description{
Generate an overlaid image of a mask and a base image
}
