% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{get_classifications}
\alias{get_classifications}
\title{Title}
\usage{
get_classifications(imgs_df, preprocess_fun_mask = NULL)
}
\arguments{
\item{imgs_df}{data frame with 3 columns, file, classif and related_file}

\item{preprocess_fun_mask}{optional preprocessing function to be used on the images}
}
\value{

}
\description{
imgs shouls be a data frame of files and classifications and related images
}
\examples{
in_df <- data.frame(
  file = c("image1_class1.png", "image1_class2.png"),
  classif = c("class1", "class2"),
  related_file = "image1.png"
)
}
