% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{hough_circles_max}
\alias{hough_circles_max}
\title{Calculates a circular hough on multiple radius and draws the best predicted circle}
\usage{
hough_circles_max(
  img,
  diameters = seq(from = 21, to = 501, by = 20),
  sobel_width = 5,
  tolerance = 21
)
}
\arguments{
\item{img}{image to transform}

\item{diameters}{diameters to go though}

\item{sobel_width}{width of the sobel filter that will be used to detect the edges}

\item{tolerance}{the position tolerance to be used during the transform}
}
\value{
image
}
\description{
This transform accentuates circular shapes
}
\examples{
x = readImage(system.file('images', 'shapes.png', package='EBImage'))
hough <- hough_circles_max(x, seq(from = 5, to = 101, by = 2), 3, 1)
display(hough, "raster")
display(x, "raster")
}
\seealso{
Other filters: 
\code{\link{circular_hough_transform}()},
\code{\link{compile_hough_circle_draw}()},
\code{\link{compile_sobel_filter}()},
\code{\link{compile_variance_filter}()},
\code{\link{dog_filter}()}
}
\concept{filters}
