% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_odd}
\alias{make_odd}
\title{Makes sure a number is odd}
\usage{
make_odd(num)
}
\arguments{
\item{num}{numeric vector}
}
\value{

}
\description{
Since many of the 2d filtering functions rely on having odd numbers for the
kernell sizes, it makes sense to have a function that assures numbers being
odd, this functions gets a number, makes it a integer and adds 1 if the
number is even, then removes duplicates.
}
\examples{
make_odd(1:20)
}
