% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{predict_img}
\alias{predict_img}
\alias{predict_img.glm}
\alias{predict_img.ranger}
\alias{predict_img.ksvm}
\alias{predict_img.default}
\alias{predict_img.glmnet}
\title{predict_img: Classify pixels based on models and features}
\usage{
predict_img(x, ...)

\method{predict_img}{glm}(x, feature_frame, ...)

\method{predict_img}{ranger}(x, feature_frame, ...)

\method{predict_img}{ksvm}(x, feature_frame, ...)

\method{predict_img}{default}(x, feature_frame, ...)

\method{predict_img}{glmnet}(x, feature_frame, ...)
}
\arguments{
\item{x}{model to use for classification}

\item{...}{Additional arguments, passed to `predict`}

\item{feature_frame}{a data frame with the features}
}
\value{
\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}
}
\description{
A wrapper that handles classifier models to output images
}
\details{
It is a fairly thin wrapper arround `predict`.
}
\examples{

}
