% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{readImageBw}
\alias{readImageBw}
\title{Read 2d images}
\usage{
readImageBw(path)
}
\arguments{
\item{path}{file path to read}
}
\value{

}
\description{
Similar to EBImage::readImage but forces it to be black and white by adding
all the channels and dividing them by the number of channels
}
\examples{

}
