% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{compile_sobel_filter}
\alias{compile_sobel_filter}
\alias{sobel_filter}
\title{Calculating sobel filters}
\usage{
compile_sobel_filter(width, dim_img)

sobel_filter(img, width = 3, img_fft = fftwtools::fftw2d(img))
}
\arguments{
\item{width}{width of the filter to be applied}

\item{dim_img}{the dimensions of the image to be used later during filtering,
only needed if pre-compiling the filter.}

\item{img}{Dimensions of the image}

\item{img_fft}{optional argument to pass the pre-computed fourier tranform of
the image from `fftwtools::fftw2d(img)`.}
}
\value{

}
\description{
The sobel filter is used to detect edges in the images
}
\section{Functions}{
\itemize{
\item \code{compile_sobel_filter}: pre-calculates parameters and returns a function
}}

\examples{
test_image <- matrix(runif(81), 9)
sobel_filter(test_image, 3)

my_sobel_filter <- compile_sobel_filter(3, dim(test_image))
my_sobel_filter(test_image)
}
\seealso{
Other filters: 
\code{\link{circular_hough_transform}()},
\code{\link{compile_hough_circle_draw}()},
\code{\link{compile_variance_filter}()},
\code{\link{dog_filter}()}
}
\concept{filters}
