% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{compile_variance_filter}
\alias{compile_variance_filter}
\alias{variance_filter}
\title{Calculate variance filter of an image}
\usage{
compile_variance_filter(width, dim_img)

variance_filter(img, width, img_fft = fftwtools::fftw2d(img))
}
\arguments{
\item{width}{width of the filter to be applied}

\item{dim_img}{the dimensions of the image to be used later during filtering,
only needed if pre-compiling the filter.}

\item{img}{Dimensions of the image}

\item{img_fft}{optional argument to pass the pre-computed fourier tranform of
the image from `fftwtools::fftw2d(img)`.}
}
\value{

}
\description{
Ths filter accentuates areas of high textre
}
\section{Functions}{
\itemize{
\item \code{compile_variance_filter}: pre-calculates parameters and returns a function
}}

\examples{
test_image <- matrix(runif(81), 9)
variance_filter(test_image, 3)

my_variance_filter <- compile_variance_filter(3, dim(test_image))
my_variance_filter(test_image)
}
\seealso{
Other filters: 
\code{\link{circular_hough_transform}()},
\code{\link{compile_hough_circle_draw}()},
\code{\link{compile_sobel_filter}()},
\code{\link{dog_filter}()},
\code{\link{hough_circles_max}()}
}
\concept{filters}
