% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{calc_features}
\alias{calc_features}
\title{Calculate filter features}
\usage{
calc_features(img, filter_widths = c(3, 5, 11, 23))
}
\arguments{
\item{filter_widths}{}
}
\value{
data.frame
}
\description{
Calculates features for each pixel based on sobel filters, gaussian and
difference of gaussians, adds as well the position in the image
}
