% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_features.R
\name{calc_features}
\alias{calc_features}
\alias{compile_calc_features}
\title{Calculate filter features}
\usage{
calc_features(
  img,
  filter_widths = c(3, 5, 11, 23),
  shape_sizes = seq(from = min(dim(img))/10, to = min(dim(img)) * 0.7, length.out = 3),
  verbose = FALSE
)

compile_calc_features(
  filter_widths = c(3, 5, 11, 23),
  shape_sizes = seq(from = min(img_dim)/10, to = min(img_dim) * 0.75, length.out = 3),
  img_dim,
  verbose = FALSE
)
}
\arguments{
\item{img}{an imput image or matrix}

\item{filter_widths}{a numeric vector of odd numbers to be used as the
width of the feature filters}

\item{shape_sizes}{a numeric vector of odd numbers indicating the expected
size of the shapes to be found}

\item{verbose}{wether to display progress messages}
}
\value{
data.frame
}
\description{
Calculates features for each pixel based on sobel filters, gaussian and
difference of gaussians, adds as well the position in the image
}
\section{Functions}{
\itemize{
\item \code{compile_calc_features}: returns a named list of functions that can be applied to an image
}}

\examples{
test_image <- matrix(runif(10201), 101)
feature_df <- calc_features(
    test_image, filter_widths = c(3,5),
    shape_sizes = c(11, 21))
head(feature_df)

feature_funs <- compile_calc_features(
     filter_widths = c(3,5), shape_sizes = c(11, 21),
     dim(test_image))
feature_funs[[1]](test_image)

feature_df2 <- purrr::map_dfc(feature_funs, ~ as.numeric(.x(test_image)))
head(feature_df2)

}
