% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{correct_light}
\alias{correct_light}
\title{Corrects uniform changes in light}
\usage{
correct_light(img, chunk_width = 200)
}
\arguments{
\item{chunk_width}{}
}
\value{

}
\description{
Corrects uniform changes in light
}
\examples{

}
