% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superpixels.R
\name{slic}
\alias{slic}
\title{Generate  superpixel mapping}
\usage{
slic(img, num_superpixel = 200, compactness = 1)
}
\arguments{
\item{img}{input image file}

\item{num_superpixel}{number of superpixels}

\item{compactness}{parameter that denotes how similarly related the pixels must be}
}
\value{
list of 2 images, one with the averaged colors and one with the indexes
}
\description{
Performs a variation of kmeans to make virtual pixels.
}
\examples{
x = readImage(system.file('images', 'shapes.png', package='EBImage'))
superpixelated <- slic(x)

# # To display the average color of each superpixel
display(superpixelated[[1]], "raster")
# # To display a colored version that makes evident the superpixel separations
display(EBImage::colorLabels(superpixelated[[2]]), "raster")
}
